+++ internal/cmd/iam/key.go
@@ -1,4 +1,4 @@
-package access
+package iam
 
 import (
 	"github.com/spf13/cobra"
